+++ src/ng-client.js
@@ -6,9 +6,14 @@ async function main() {
   const options = {
     address: argv.a || "localhost",
     port: argv.p || 2113,
+    anonymous: argv.n || false,
   };
 
   const args = ["-f", "json", "-i"];
+  if (options.anonymous) {
+    args.push("-a");
+    delete options.anonymous;
+  }
 
   const nail = nailgunClient.exec("net.dangmai.serializer.Apex", args, options);