+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Add support for Anonymous Code block with `--apex-anonymous` option.
 - CLI/Option change:
   - `use-standalone-server` option is now `apex-standalone-parser`,
   and it is now a choice between `none` and `built-in`.